+++ ttcr/ctpl_stl.h
@@ -73,7 +73,7 @@ namespace ctpl {
     public:
 
         thread_pool() { this->init(); }
-        thread_pool(int nThreads) { this->init(); this->resize(nThreads); }
+        thread_pool(size_t nThreads) { this->init(); this->resize(nThreads); }
 
         // the destructor waits for all the functions in the queue to be finished
         ~thread_pool() {
@@ -81,29 +81,29 @@ namespace ctpl {
         }
 
         // get the number of running threads in the pool
-        int size() { return static_cast<int>(this->threads.size()); }
+        size_t size() { return this->threads.size(); }
 
         // number of idle threads
-        int n_idle() { return this->nWaiting; }
-        std::thread & get_thread(int i) { return *this->threads[i]; }
+        size_t n_idle() { return this->nWaiting; }
+        std::thread & get_thread(size_t i) { return *this->threads[i]; }
 
         // change the number of threads in the pool
         // should be called from one thread, otherwise be careful to not interleave, also with this->stop()
         // nThreads must be >= 0
-        void resize(int nThreads) {
+        void resize(size_t nThreads) {
             if (!this->isStop && !this->isDone) {
-                int oldNThreads = static_cast<int>(this->threads.size());
+                size_t oldNThreads = this->threads.size();
                 if (oldNThreads <= nThreads) {  // if the number of threads is increased
                     this->threads.resize(nThreads);
                     this->flags.resize(nThreads);
 
-                    for (int i = oldNThreads; i < nThreads; ++i) {
+                    for (size_t i = oldNThreads; i < nThreads; ++i) {
                         this->flags[i] = std::make_shared<std::atomic<bool>>(false);
                         this->set_thread(i);
                     }
                 }
                 else {  // the number of threads is decreased
-                    for (int i = oldNThreads - 1; i >= nThreads; --i) {
+                    for (size_t i = oldNThreads - 1; i >= nThreads; --i) {
                         *this->flags[i] = true;  // this thread will finish
                         this->threads[i]->detach();
                     }
@@ -120,17 +120,17 @@ namespace ctpl {
 
         // empty the queue
         void clear_queue() {
-            std::function<void(int id)> * _f;
+            std::function<void(size_t id)> * _f;
             while (this->q.pop(_f))
                 delete _f; // empty the queue
         }
 
         // pops a functional wrapper to the original function
-        std::function<void(int)> pop() {
-            std::function<void(int id)> * _f = nullptr;
+        std::function<void(size_t)> pop() {
+            std::function<void(size_t id)> * _f = nullptr;
             this->q.pop(_f);
-            std::unique_ptr<std::function<void(int id)>> func(_f); // at return, delete the function even if an exception occurred
-            std::function<void(int)> f;
+            std::unique_ptr<std::function<void(size_t id)>> func(_f); // at return, delete the function even if an exception occurred
+            std::function<void(size_t)> f;
             if (_f)
                 f = *_f;
             return f;
@@ -144,7 +144,7 @@ namespace ctpl {
                 if (this->isStop)
                     return;
                 this->isStop = true;
-                for (int i = 0, n = this->size(); i < n; ++i) {
+                for (size_t i = 0, n = this->size(); i < n; ++i) {
                     *this->flags[i] = true;  // command the threads to stop
                 }
                 this->clear_queue();  // empty the queue
@@ -158,7 +158,7 @@ namespace ctpl {
                 std::unique_lock<std::mutex> lock(this->mutex);
                 this->cv.notify_all();  // stop all waiting threads
             }
-            for (int i = 0; i < static_cast<int>(this->threads.size()); ++i) {  // wait for the computing threads to finish
+            for (size_t i = 0; i < static_cast<size_t>(this->threads.size()); ++i) {  // wait for the computing threads to finish
                 if (this->threads[i]->joinable())
                     this->threads[i]->join();
             }
@@ -171,10 +171,10 @@ namespace ctpl {
 
         template<typename F, typename... Rest>
         auto push(F && f, Rest&&... rest) ->std::future<decltype(f(0, rest...))> {
-            auto pck = std::make_shared<std::packaged_task<decltype(f(0, rest...))(int)>>(
+            auto pck = std::make_shared<std::packaged_task<decltype(f(0, rest...))(size_t)>>(
                                                                                           std::bind(std::forward<F>(f), std::placeholders::_1, std::forward<Rest>(rest)...)
                                                                                           );
-            auto _f = new std::function<void(int id)>([pck](int id) {
+            auto _f = new std::function<void(size_t id)>([pck](size_t id) {
                 (*pck)(id);
             });
             this->q.push(_f);
@@ -187,8 +187,8 @@ namespace ctpl {
         // operator returns std::future, where the user can get the result and rethrow the catched exceptins
         template<typename F>
         auto push(F && f) ->std::future<decltype(f(0))> {
-            auto pck = std::make_shared<std::packaged_task<decltype(f(0))(int)>>(std::forward<F>(f));
-            auto _f = new std::function<void(int id)>([pck](int id) {
+            auto pck = std::make_shared<std::packaged_task<decltype(f(0))(size_t)>>(std::forward<F>(f));
+            auto _f = new std::function<void(size_t id)>([pck](size_t id) {
                 (*pck)(id);
             });
             this->q.push(_f);
@@ -206,15 +206,15 @@ namespace ctpl {
         thread_pool & operator=(const thread_pool &);// = delete;
         thread_pool & operator=(thread_pool &&);// = delete;
 
-        void set_thread(int i) {
+        void set_thread(size_t i) {
             std::shared_ptr<std::atomic<bool>> flag(this->flags[i]); // a copy of the shared ptr to the flag
             auto f = [this, i, flag/* a copy of the shared ptr to the flag */]() {
                 std::atomic<bool> & _flag = *flag;
-                std::function<void(int id)> * _f;
+                std::function<void(size_t id)> * _f;
                 bool isPop = this->q.pop(_f);
                 while (true) {
                     while (isPop) {  // if there is anything in the queue
-                        std::unique_ptr<std::function<void(int id)>> func(_f); // at return, delete the function even if an exception occurred
+                        std::unique_ptr<std::function<void(size_t id)>> func(_f); // at return, delete the function even if an exception occurred
                         (*_f)(i);
                         if (_flag)
                             return;  // the thread is wanted to stop, return even if the queue is not empty yet
@@ -237,10 +237,10 @@ namespace ctpl {
 
         std::vector<std::unique_ptr<std::thread>> threads;
         std::vector<std::shared_ptr<std::atomic<bool>>> flags;
-        detail::Queue<std::function<void(int id)> *> q;
+        detail::Queue<std::function<void(size_t id)> *> q;
         std::atomic<bool> isDone;
         std::atomic<bool> isStop;
-        std::atomic<int> nWaiting;  // how many threads are waiting
+        std::atomic<size_t> nWaiting;  // how many threads are waiting
 
         std::mutex mutex;
         std::condition_variable cv;
